fix(calls): move call alert mute toggle to mentions tab
Remove the bell toggle from the call join strip so the banner wraps cleanly, and add a volume mute/unmute control in the mentions inbox header instead.
